+++ README.md
@@ -2,4 +2,4 @@
 
 #Installation 
 
-1. Simply run the following command `pip install -r requirements.txt ` in the terminal.
+1. Simply run the following command `pip install -r requirements.txt` in the terminal.
