+++ mask/masking_plugins/Mask_abstract.py
