+++ mask/masking_plugins/Mask_names_simple.py
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 # Code by: Nikola Milosevic
-from masking_plugins.Mask_abstract import Mask_abstract
+from mask.masking_plugins.Mask_abstract import Mask_abstract
 import random
 
 class Mask_names_simple(Mask_abstract):
