+++ mask/masking_plugins/mask_redact.py
@@ -1,4 +1,4 @@
-from masking_plugins.Mask_abstract import Mask_abstract
+from mask.masking_plugins.Mask_abstract import Mask_abstract
 
 
 class MaskRedact(Mask_abstract):
