+++ mask/source/configuration.cnf
